Use RAII wrapper to manage OpenSSL cipher contexts and cipher implementations
The resources are now cleaned up automatically whenever the CryptoCipher object goes out of scope, such as in case of an error, or when the functions return. The previous implementation depended on std::unique_ptr, but I think  encapsulating the data and operations on a cipher into a class would be more helpful and would provide an interface for easier handling of the cipher.