Use BLAKE2b on 64-bit arch, and BLAKE2s on lower archs
BLAKE2b is optimized for 64-bit architectures while BLAKE2s is optimized for lower architectures. The aim of the code is to identify duplicates as fast as possible, as the user might scan a directory with a large number of files.
The future version aims to use BLAKE3, which is just one algorithm optimized for all architectures, is much faster than BLAKE2, highly parallelizable and secure. BLAKE3 is not currently implemented in most cryptographic libraries, but an official implementation exists. The C version does not support parallelization yet, so I would probably consider going for it, or call the parallelized Rust version directly from C++.