BugFix: Template instantiation issues
The 'printColor(...)' function has been moved to the 'utility.hpp' header to solve issues with template instantiation. The constraining concept for the templated type must be defined in the same translation unit as the function using it, and since the function should be accessible to other translation units in the project, both (the concept and the function) have to be defined in a header.
Minor improvements (and verbosity) for 'getHomeDir()' have been added as well.