Integrate encryption/decryption concurrency into the password manager
The password field of a password record is now encrypted twice, first with Serpent, and then AES. The other fields, site and username are encrypted once with AES for storage.
In the memory at runtime, the site and username fields are not encrypted, but the password fields are encrypted with Serpent, unless an operation requires them unencrypted, after which they are re-encrypted.