Refactor BLAKE3 build step in GitHub Workflow
The workflow file 'cmake-multi-platform.yml' for multi-platform CMake builds has been modified to simplify and improve the BLAKE3 installation process. Instead of directly cloning and building BLAKE3 inside the job, an install script was incorporated which is called based on the operating system detected. This implementation ensures better OS compatibility and enhances the overall clarity and maintainability of the code.